chore: rename app endpoint
